requirements.txt
codegen==1.0
Django==1.10.1
django-filter==0.15.3
djangorestframework==3.5.1
freeze==1.0.10
Markdown==2.6.7
patterns==0.3
six==1.10.0
